Fix CHECK_EQ type mismatch
Cast to int32_t required as MediaPuller enum uses multicharacter enum values
which may not be properly handled by the CHECK_EQ macro.

+++ media/libstagefright/wifi-display/source/Converter.cpp
@@ -340,7 +340,8 @@ void Converter::onMessageReceived(const sp<AMessage> &msg) {
 
                 scheduleDoMoreWork();
             } else {
-                CHECK_EQ(what, MediaPuller::kWhatAccessUnit);
+                CHECK_EQ(what, static_cast<int32_t>(
+                    MediaPuller::kWhatAccessUnit));
 
                 sp<ABuffer> accessUnit;
                 CHECK(msg->findBuffer("accessUnit", &accessUnit));
